Refactor ContactConfigScreen.kt
